Move "Kernel\TestKernel" to the namespace root
