+++ composer.json
@@ -27,6 +27,7 @@
     "symfony/config": "^5.4 || ^6.4",
     "symfony/console": "^5.4 || ^6.4",
     "symfony/dependency-injection": "^5.4 || ^6.4",
+    "symfony/deprecation-contracts": "^2.1 || ^3.0",
     "symfony/event-dispatcher": "^5.4 || ^6.4",
     "symfony/filesystem": "^5.4 || ^6.4",
     "symfony/framework-bundle": "^5.4 || ^6.4",
