+++ phpstan.neon
@@ -11,6 +11,8 @@ parameters:
 	excludePaths:
 		# There are two errors, that were not able to be added to the baseline
 		- src/Kernel/CompatibilityKernel.php
+		# Deprecated aliases
+		- src/Kernel/TestKernel.php
 
 	bootstrapFiles:
 		- vendor/pimcore/pimcore/stubs/dynamic-constants.php
