+++ tests/Functional/KernelShutdownTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Neusta\Pimcore\TestingFramework\Tests\Functional;
 
-use Neusta\Pimcore\TestingFramework\Kernel\TestKernel;
 use Neusta\Pimcore\TestingFramework\Test\ConfigurableKernelTestCase;
+use Neusta\Pimcore\TestingFramework\TestKernel;
 use Symfony\Component\Filesystem\Filesystem;
 
 class KernelShutdownTest extends ConfigurableKernelTestCase
